The app fails in section 6 without some changes (#114)
* The app fails in section 6 without some changes

In section 6, checking for changes will cause all kinds of problems unless the setCustomer(null); is commented out. 

I did not add a section where the reader needs to uncomment this, with the assumption that this is clear.

* Changed the location of the changes

Removed the old commit changes, and replaced them with needed changes at a more appropriate location. No commenting used, but rather delayed the addition of setCustomer(null) to a later point.

* Specified the setter method

Specified that the setter method in question is setCustomer.

* Add backticks around setCustomer

Also rephrased the sentence a bit.